Refactor to Query Helper, UserQuery Completed Created a Query Helper module to help dry up queries and ease caching implementation
Refactor LanguageQuery

Refactors FollowerQuery

Refactor CommitTimeline

Refactor Commit Message
