+++ app/queries/follower_query.rb
@@ -1,14 +1,11 @@
 class FollowerQuery
+  include QueryHelper
 
   def self.execute_query(username)
     new.query(username)
   end
 
-  def query(username)
-   validate_response(username)
-  end
-
-  def commit_query(username)
+  def query_maker(username)
     github_service(GQLi::DSL.query {
                      user(login: username) {
                        followers(first: 100) {
@@ -19,18 +16,4 @@ class FollowerQuery
     })
   end
 
-  def validate_response(username)
-    response = commit_query(username)
-    return build_failure_response(username) unless response["user"]
-    return response["user"]
-  end
-
-
-  def github_service(query_data)
-    GithubApiInterface.get(query_data)
-  end
-
-  def build_failure_response(username)
-    {error:  "User #{username} Not Found"}
-  end
 end
