+++ app/queries/language_query.rb
@@ -1,21 +1,13 @@
 class LanguageQuery
+  include QueryHelper
+
   def self.execute_query(username)
     new.query(username)
   end
 
-  def query(username)
-    validate_response(username)
-  end
-
   private
 
-  def validate_response(username)
-    response = make_query(username)
-    return build_failure_response(username) if response["user"].nil?
-    return response["user"]
-  end
-
-  def make_query(username)
+  def query_maker(username)
     github_service(GQLi::DSL.query {
       user(login: username) {
         repositories(last: 100) {
@@ -33,13 +25,5 @@ class LanguageQuery
         }
       }
     })
-  end
-
-  def github_service(query_data)
-    GithubApiInterface.get(query_data)
-  end
-
-  def build_failure_response(username)
-    {error: "User #{username} Not Found"}
   end
 end
